Did some refactoring on PostEdit and CreatePost activities, consolidating composables to reduce unnecessary redundancies.
