+++ app/src/main/java/com/jerboa/ui/components/post/create/CreatePostActivity.kt
@@ -1,5 +1,6 @@
 package com.jerboa.ui.components.post.create
 
+import android.annotation.SuppressLint
 import android.net.Uri
 import android.util.Log
 import android.util.Patterns
@@ -15,18 +16,27 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.navigation.NavController
 import com.jerboa.DEBOUNCE_DELAY
+import com.jerboa.R
 import com.jerboa.api.ApiState
 import com.jerboa.api.uploadPictrsImage
 import com.jerboa.datatypes.types.CreatePost
 import com.jerboa.datatypes.types.GetSiteMetadata
+import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
 import com.jerboa.imageInputStreamFromUri
 import com.jerboa.ui.components.common.LoadingBar
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.community.list.CommunityListViewModel
+import com.jerboa.ui.components.post.composables.CreateEditPostBody
+import com.jerboa.ui.components.post.composables.CreateEditPostHeader
+import com.jerboa.ui.components.post.composables.CreatePostSubmitIcon
+import com.jerboa.ui.components.post.composables.PostCommunitySelector
+import com.jerboa.validatePostName
+import com.jerboa.validateUrl
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
@@ -35,6 +45,7 @@ private var fetchSiteMetadataJob: Job? = null
 
 data class MetaDataRes(val title: String?, val loading: Boolean)
 
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter", "UnrememberedMutableState")
 @Composable
 fun CreatePostActivity(
     accountViewModel: AccountViewModel,
@@ -50,9 +61,11 @@ fun CreatePostActivity(
     val ctx = LocalContext.current
     val account = getCurrentAccount(accountViewModel = accountViewModel)
     val scope = rememberCoroutineScope()
+    val community = communityListViewModel.selectedCommunity
 
     var name by rememberSaveable { mutableStateOf("") }
     var url by rememberSaveable { mutableStateOf(initialUrl) }
+    var isNsfw by rememberSaveable { mutableStateOf(false) }
     var body by rememberSaveable(stateSaver = TextFieldValue.Saver) {
         mutableStateOf(
             TextFieldValue(
@@ -60,8 +73,10 @@ fun CreatePostActivity(
             ),
         )
     }
-    var isNsfw by rememberSaveable { mutableStateOf(false) }
-    var formValid by rememberSaveable { mutableStateOf(false) }
+
+    val nameField = validatePostName(name)
+    val urlField = validateUrl(url)
+    val formValid = !nameField.hasError && !urlField.hasError && (community !== null)
 
     LaunchedEffect(initialUrl) {
         if (initialUrl.isNotEmpty()) {
@@ -90,44 +105,43 @@ fun CreatePostActivity(
                     else -> false
                 }
                 Column {
-                    CreatePostHeader(
+                    CreateEditPostHeader(
                         navController = navController,
                         formValid = formValid,
                         loading = loading,
-                        onCreatePostClick = {
-                            account?.also { acct ->
-                                communityListViewModel.selectedCommunity?.id?.also {
-                                    // Clean up that data
-                                    val nameOut = name.trim()
-                                    val bodyOut = body.text.trim().ifEmpty { null }
-                                    val urlOut = url.trim().ifEmpty { null }
-                                    createPostViewModel.createPost(
-                                        CreatePost(
-                                            name = nameOut,
-                                            community_id = it,
-                                            url = urlOut,
-                                            body = bodyOut,
-                                            auth = acct.jwt,
-                                            nsfw = isNsfw,
-                                        ),
-                                        navController,
-                                    )
-                                }
-                            }
+                        onSubmitClick = {
+                            onSubmitClick(
+                                name,
+                                body,
+                                url,
+                                isNsfw,
+                                account,
+                                communityListViewModel,
+                                createPostViewModel,
+                                navController,
+                            )
                         },
+                        submitIcon = {
+                            CreatePostSubmitIcon(formValid)
+                        },
+                        title = stringResource(R.string.create_post_create_post),
                     )
+
                     if (loading) {
                         LoadingBar()
                     }
                 }
             },
             content = { padding ->
-                CreatePostBody(
+
+                CreateEditPostBody(
                     name = name,
+                    nameField = nameField,
                     onNameChange = { name = it },
                     body = body,
                     onBodyChange = { body = it },
                     url = url,
+                    urlField = urlField,
                     onUrlChange = { cUrl ->
                         url = cUrl
                         fetchSiteMetadataJob?.cancel()
@@ -138,9 +152,6 @@ fun CreatePostActivity(
                             }
                         }
                     },
-                    navController = navController,
-                    community = communityListViewModel.selectedCommunity,
-                    formValid = { formValid = it },
                     suggestedTitle = suggestedTitle,
                     suggestedTitleLoading = suggestedTitleLoading,
                     image = initialImage,
@@ -158,8 +169,45 @@ fun CreatePostActivity(
                     padding = padding,
                     isNsfw = isNsfw,
                     onIsNsfwChange = { isNsfw = it },
+                    communitySelector = {
+                        PostCommunitySelector(
+                            community = community,
+                            navController = navController,
+                        )
+                    },
                 )
             },
         )
+    }
+}
+
+fun onSubmitClick(
+    name: String,
+    body: TextFieldValue,
+    url: String,
+    isNsfw: Boolean,
+    account: Account?,
+    communityListViewModel: CommunityListViewModel,
+    createPostViewModel: CreatePostViewModel,
+    navController: NavController,
+) {
+    account?.also { acct ->
+        communityListViewModel.selectedCommunity?.id?.also {
+            // Clean up that data
+            val nameOut = name.trim()
+            val bodyOut = body.text.trim().ifEmpty { null }
+            val urlOut = url.trim().ifEmpty { null }
+            createPostViewModel.createPost(
+                CreatePost(
+                    name = nameOut,
+                    community_id = it,
+                    url = urlOut,
+                    body = bodyOut,
+                    auth = acct.jwt,
+                    nsfw = isNsfw,
+                ),
+                navController,
+            )
+        }
     }
 }
