+++ app/src/main/java/com/jerboa/ui/components/post/edit/PostEditActivity.kt
@@ -1,11 +1,10 @@
 
 package com.jerboa.ui.components.post.edit
 
+import android.annotation.SuppressLint
 import android.util.Log
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -13,13 +12,15 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.navigation.NavController
+import com.jerboa.R
 import com.jerboa.api.ApiState
 import com.jerboa.api.uploadPictrsImage
 import com.jerboa.datatypes.types.EditPost
+import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
 import com.jerboa.imageInputStreamFromUri
 import com.jerboa.ui.components.common.LoadingBar
@@ -28,9 +29,14 @@ import com.jerboa.ui.components.community.CommunityViewModel
 import com.jerboa.ui.components.home.HomeViewModel
 import com.jerboa.ui.components.person.PersonProfileViewModel
 import com.jerboa.ui.components.post.PostViewModel
+import com.jerboa.ui.components.post.composables.CreateEditPostBody
+import com.jerboa.ui.components.post.composables.CreateEditPostHeader
+import com.jerboa.ui.components.post.composables.EditPostSubmitIcon
+import com.jerboa.validatePostName
+import com.jerboa.validateUrl
 import kotlinx.coroutines.launch
 
-@OptIn(ExperimentalMaterial3Api::class)
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
 fun PostEditActivity(
     accountViewModel: AccountViewModel,
@@ -46,8 +52,8 @@ fun PostEditActivity(
     val ctx = LocalContext.current
     val account = getCurrentAccount(accountViewModel = accountViewModel)
     val scope = rememberCoroutineScope()
-
     val pv = postEditViewModel.postView
+
     var name by rememberSaveable { mutableStateOf(pv?.post?.name.orEmpty()) }
     var url by rememberSaveable { mutableStateOf(pv?.post?.url.orEmpty()) }
     var isNsfw by rememberSaveable { mutableStateOf(pv?.post?.nsfw ?: false) }
@@ -58,7 +64,10 @@ fun PostEditActivity(
             ),
         )
     }
-    var formValid by rememberSaveable { mutableStateOf(true) }
+
+    val nameField = validatePostName(name)
+    val urlField = validateUrl(url)
+    val formValid = !nameField.hasError && !urlField.hasError
 
     Scaffold(
         topBar = {
@@ -67,34 +76,30 @@ fun PostEditActivity(
                     ApiState.Loading -> true
                     else -> false
                 }
-                EditPostHeader(
+                CreateEditPostHeader(
                     navController = navController,
                     formValid = formValid,
                     loading = loading,
-                    onEditPostClick = {
-                        account?.also { acct ->
-                            // Clean up that data
-                            val nameOut = name.trim()
-                            val bodyOut = body.text.trim().ifEmpty { null }
-                            val urlOut = url.trim().ifEmpty { null }
-
-                            postEditViewModel.editPost(
-                                form = EditPost(
-                                    post_id = pv!!.post.id,
-                                    name = nameOut,
-                                    url = urlOut,
-                                    body = bodyOut,
-                                    auth = acct.jwt,
-                                    nsfw = isNsfw,
-                                ),
-                                navController = navController,
-                                postViewModel = postViewModel,
-                                personProfileViewModel = personProfileViewModel,
-                                communityViewModel = communityViewModel,
-                                homeViewModel = homeViewModel,
-                            )
-                        }
+                    onSubmitClick = {
+                        onSubmitClick(
+                            account = account,
+                            name = name,
+                            body = body,
+                            url = url,
+                            postEditViewModel = postEditViewModel,
+                            isNsfw = isNsfw,
+                            navController = navController,
+                            postViewModel = postViewModel,
+                            communityViewModel = communityViewModel,
+                            personProfileViewModel = personProfileViewModel,
+                            homeViewModel = homeViewModel,
+                        )
                     },
+                    submitIcon = {
+                        EditPostSubmitIcon()
+                    },
+                    title = stringResource(R.string.post_edit_edit_post),
+
                 )
                 if (loading) {
                     LoadingBar()
@@ -102,14 +107,15 @@ fun PostEditActivity(
             }
         },
         content = { padding ->
-            EditPostBody(
+            CreateEditPostBody(
                 name = name,
+                nameField = nameField,
                 onNameChange = { name = it },
                 body = body,
                 onBodyChange = { body = it },
                 url = url,
+                urlField = urlField,
                 onUrlChange = { url = it },
-                formValid = { formValid = it },
                 onPickedImage = { uri ->
                     val imageIs = imageInputStreamFromUri(ctx, uri)
                     scope.launch {
@@ -119,12 +125,49 @@ fun PostEditActivity(
                     }
                 },
                 account = account,
-                modifier = Modifier
-                    .padding(padding)
-                    .imePadding(),
+                padding = padding,
                 isNsfw = isNsfw,
                 onIsNsfwChange = { isNsfw = it },
+                communitySelector = {},
             )
         },
     )
+}
+
+fun onSubmitClick(
+    account: Account?,
+    name: String,
+    body: TextFieldValue,
+    url: String,
+    postEditViewModel: PostEditViewModel,
+    isNsfw: Boolean,
+    navController: NavController,
+    postViewModel: PostViewModel,
+    communityViewModel: CommunityViewModel,
+    personProfileViewModel: PersonProfileViewModel,
+    homeViewModel: HomeViewModel,
+) {
+    account?.also { acct ->
+        // Clean up that data
+        val nameOut = name.trim()
+        val bodyOut = body.text.trim().ifEmpty { null }
+        val urlOut = url.trim().ifEmpty { null }
+        val pv = postEditViewModel.postView
+
+        postEditViewModel.editPost(
+            form = EditPost(
+                post_id = pv!!.post.id,
+                name = nameOut,
+                url = urlOut,
+                body = bodyOut,
+                auth = acct.jwt,
+                nsfw = isNsfw,
+            ),
+            navController = navController,
+            postViewModel = postViewModel,
+            personProfileViewModel = personProfileViewModel,
+            communityViewModel = communityViewModel,
+            homeViewModel = homeViewModel,
+        )
+    }
 }
